Upgrade workflows to `gradle-build-action@v2`
Upgrades the GitHub actions workflows to use `gradle/gradle-build-action@v2` (presently v2.0-beta.7).
This action implementation replaces the previous `eskatos/gradle-command-action@v1`.

This new action:
- has simpler configuration
- does a better job of caching the Gradle User Home state between build invocations
- automatically adds annotations with build scan URLs for all gradle invocations.

Due to the second item, I've removed the separate use of `actions/cache` for caching.
I'd be very interested to know if there is any regression in performance with this change.